refactor(reaction): totalCount 타입 변경 Long -> Int

+++ src/main/kotlin/kr/co/wground/reaction/application/ReactionQueryService.kt
@@ -37,7 +37,7 @@ class ReactionQueryService(
 
         return PostReactionStats(
             postId = postId,
-            totalCount = reactions.size.toLong(),
+            totalCount = reactions.size,
             summaries = summaries
         )
     }
